requirements.txt
openai>=1.0.0
anthropic>=0.9.0
pyyaml>=6.0
requests>=2.31.0
argparse>=1.4.0
colorama>=0.4.6
ollama>=0.1.0
streamlit>=1.31.0